requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.6
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
